ml/requirements.txt
google-cloud-language>=1.3.0